+++ tests/UnitTestEx.NUnit.Test/PersonFunctionTest.cs
@@ -87,5 +87,15 @@ namespace UnitTestEx.NUnit.Test
                 .AssertOK()
                 .Assert(new { first = "Rachel", last = "Smith" });
         }
+
+        [Test]
+        public void ValueVsHttpRequestContent()
+        {
+            using var test = FunctionTester.Create<Startup>();
+            test.HttpTrigger<PersonFunction>()
+                .Run(f => f.RunWithContent(new Person { FirstName = "Rachel", LastName = "Smith" }, test.Logger))
+                .AssertOK()
+                .Assert(new { first = "Rachel", last = "Smith" });
+        }
     }
 }